// src/features/user-profile/interfaces/IFormValues.ts
export interface FormValues {
    firstname: string;
    lastname: string;
    email: string;
    dob: string;
    phone: string;
    country: string;
    institute: string;
    stream: string;
    passingYear: string;
    subjectOfInterest: string;
    preferredLanguage: string;
    accessibilityRequirement: string;
  }